# Speed up stringification