String streams continue to be slow. :)

Benchmark results from running google benchmark on the side:
```
------------------------------------------------------------
Benchmark                  Time             CPU   Iterations
------------------------------------------------------------
BM_SoleStream            526 ns          526 ns      1312873
BM_SoleIter             31.8 ns         31.8 ns     22381764
```

From benchmark added in this repo:
```
Before:
Benchmarking str... 1545601 uuids/sec
After:
Benchmarking str... 13048124 uuids/sec
```