@r-lyeh 2021-03-16
Yeah I was waiting to see if anybody would give any feedback on this. However, since the use case community is so small I think it is better to merge & fix if troubles arise :D